requirements.txt
confluent-kafka
requests
elasticsearch
kafka-python
python-dotenv
azure-storage-file-datalake